Added in more support for unit tests setup in issue #1.
Have the Makefile now building and linking the object files for creating
the unit tests. Need to see about getting the Makefile more automated
for pulling in the test files and compiling them.

Another item that needs to be sorted out is dependencies between the
Make targets: making the directory structure before building the
objects.

Adding in the first unit test files, just showing where the constructor
throws when bad parameters are passed.