Updating Makefile
Added the tests to the default target.

Integrating the build target into the test target to ensure that the
build directory is created before compilation of the tests.
Updated placement of g++ arguments in tests target to be consistent.

Inserted rule for removing the object files from the src directory.

TODO: get the object files into their own directory.