Added test case for single crowning moves.
Working on #5.

Have added in a BDD style tests to have scenarios for
crowning moves. These are just simple tests where the are 1 piece of red
and blk in each scenario. There should be no interaction between the
said peices and get one possible move from the move generator that
results in crowning a red and blk peice in each scenario respectfully.