Fixing testing of output of the board test.
If want to have the tests preformed and outputted within the contexts of
Catch, need to have the output done with Catch's messagers.

Pulls in the couts in 263d30c4f4ec9ca8fa06cdfadfc8e7123ff8fb31 to a
REQUIRE for comparing the string form of the boards against each other.